ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
langchain==0.3.0
langchain-groq==0.2.0
langchain-community==0.3.0
sentence-transformers==3.1.0
supabase==2.7.0
PyMuPDF==1.24.10
python-docx==1.1.2
pydantic==2.9.1
reportlab==4.2.2
vecs==0.4.3
numpy==1.26.4
httpx==0.27.2
